+++ eyes.appium.java/src/main/java/com/applitools/eyes/appium/locators/BaseVisualLocatorProvider.java
@@ -15,7 +15,6 @@ import java.util.Map;
 
 public abstract class BaseVisualLocatorProvider implements IVisualLocatorProvider {
 
-    private String viewportScreenshotUrl = null;
     protected Logger logger;
     private IServerConnector serverConnector;
     protected EyesAppiumDriver driver;
@@ -29,25 +28,13 @@ public abstract class BaseVisualLocatorProvider implements IVisualLocatorProvide
     }
 
     @Override
-    public void tryPostScreenshotForLocators() {
-        logger.verbose("Try to upload viewport screenshot");
-        String base64Image = ImageUtils.base64FromImage(getViewPortScreenshot());
-        if (base64Image != null) {
-            viewportScreenshotUrl = serverConnector.postViewportImage(base64Image);
-            logger.verbose("Done");
-        }
-    }
-
-    @Override
     public Map<String, List<Region>> getLocators(IVisualLocatorSettings visualLocatorSettings) {
         ArgumentGuard.notNull(visualLocatorSettings, "visualLocatorSettings");
 
         logger.verbose("Get locators with given names: " + visualLocatorSettings.getNames());
 
-        if (viewportScreenshotUrl == null) {
-            logger.verbose("Viewport screenshot was not uploaded");
-            tryPostScreenshotForLocators();
-        }
+        String base64Image = ImageUtils.base64FromImage(getViewPortScreenshot());
+        String viewportScreenshotUrl = serverConnector.postViewportImage(base64Image);
 
         VisualLocatorsData data = new VisualLocatorsData(driver.getEyes().getAppName(), viewportScreenshotUrl, visualLocatorSettings.isFirstOnly(), visualLocatorSettings.getNames());